public/files/week10/week10-1-elasticache-lab/elasticache-lab/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
PyMySQL==1.1.0
python-dotenv==1.0.0
pydantic==2.5.3
